Refactorize the program
Modify the govet function to return the go vet diagnostic message and a
non nil error only in case of a fatal problem.  This change greatly
simplifies the code.

Rename the syntax function to run, and change it so that it is only
responsible for printing the go vet diagnostic messages.